+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format of this change log follows the advice given at [Keep a CHANGELOG](htt
 
 ## [Unreleased]
 
+### Added
+
+- Support for `moodle-package-` prefix which is not a moodle plugintype.
+
 ### Fixed
 
 - Fix linting issues.